+++ lib/flutter_openim_sdk.dart
@@ -13,6 +13,7 @@ export 'src/enum/sdk_error_code.dart';
 export 'src/listener/advanced_msg_listener.dart';
 export 'src/listener/connect_listener.dart';
 export 'src/listener/conversation_listener.dart';
+export 'src/listener/custom_business_listener.dart';
 export 'src/listener/friendship_listener.dart';
 export 'src/listener/group_listener.dart';
 export 'src/listener/msg_send_progress_listener.dart';