+++ lib/src/listener/custom_business_listener.dart
@@ -1,0 +1,9 @@
+class OnCustomBusinessListener {
+  Function(String s)? onRecvCustomBusinessMessage;
+
+  OnCustomBusinessListener({this.onRecvCustomBusinessMessage});
+
+  void recvCustomBusinessMessage(String s) {
+    onRecvCustomBusinessMessage?.call(s);
+  }
+}